don't autodetect hi-/lo-DPI: add/remove/modify F scripts
Automatically detecting whether a screen is low or high DPI is
tricky on a multi-monitor setup. Instead, make the user specify
whether the screen is low- or high-DPI.

F is modified accordingly and now supports relative increase
and decrease of font size instead of setting a new font and size
directly.

F+ and F- are modified to always assume a low-DPI screen.

F++ and F-- are added, which are like F+ and F- but assume a
high-DPI screen.

ishidpi is removed as it is no longer needed.